move spline4d source into spline.cpp
